Fix to execute user-creating fixture properly
